frontend: Upgrade storybook setup
This is a pretty big commit that upgrades storybook version and also
changes the way we mock backend in stories.

Previosly stories would override methods in particular resources
Like Pod.useList = () => mockData. But this wasn't playing nicely with
vite and wasn't reliable as we were modifying global values during
runtime.

Instead all of that mocking is happening in the mock service worker
(msw). Which intersepts network requests and is a reliable way of
mocking.

In addition to that, the storybook.test.tsx also had some changes to run
the stories in a way the documentation recommends. This also changed the
output of the snapshots slightly, which required regeneration of all
snapshots.

Signed-off-by: Oleksandr Dubenko <oldubenko@microsoft.com>